Split gridworld_examples.py into two files: algs and envs
Much cleaner this way
